update router, register component
