+++ src/pages/Register.jsx
@@ -1,13 +1,57 @@
-import React, { useState } from "react";
-import { Link } from "react-router";
+import { Link, useNavigate } from "react-router";
 import { Eye, EyeOff } from "lucide-react";
 import Animation from "../components/Animation/Animation";
+import { useForm } from "react-hook-form";
+import { useDispatch } from "react-redux";
+import { registerUser } from "../components/feature/authSlice";
+import { useState } from "react";
+import axios from "axios";
+
 
 const Register = () => {
-  const [showPassword, setShowPassword] = useState(false);
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm();
+  const [profilePic, setProfilePic] = useState("");
+  const dispatch = useDispatch()
+  const navigate = useNavigate();
 
-  const togglePasswordVisibility = () => {
-    setShowPassword(!showPassword);
+
+  const onSubmit = (data) => {
+    dispatch(
+      registerUser({
+        email: data.email,
+        password: data.password,
+        name: data.name,
+        photoURL: profilePic,
+      })
+    )
+      .unwrap()
+      .then(() => {
+        // toast.success("Successfully registered!");
+        navigate("/");
+      })
+      .catch((err) => {
+        // toast.error("Failed to register");
+        console.log(err)
+      });
+
+  }
+  const handleImageUp = async (e) => {
+    const image = e.target.files[0];
+    const formData = new FormData();
+    formData.append("image", image);
+
+    const uploadURL = `https://api.imgbb.com/1/upload?key=${import.meta.env.VITE_image_key}`;
+    try {
+      const res = await axios.post(uploadURL, formData);
+      const imageUrl = res.data?.data?.url;
+      setProfilePic(imageUrl);
+    } catch (err) {
+      console.error("Image upload failed", err);
+    }
   };
   return (
     <div className="w-11/12 mx-auto grid grid-cols-1 md:grid-cols-2 items-center justify-center gap-5 my-5">
@@ -22,76 +66,35 @@ const Register = () => {
             LogIn Here
           </Link>
         </p>
-        <form className="space-y-8">
-          <div className="space-y-4">
-            <div className="space-y-2">
-              <label htmlFor="email" className="block text-sm">
-                User Name
-              </label>
-              <input
-                type="text"
-                name="name"
-                id="name"
-                required
-                placeholder="User Name"
-                className="w-full px-3 py-2 border rounded-md border-gray-300 bg-gray-50 text-gray-800 focus:border-violet-600"
-              />
-            </div>
-            <div className="space-y-2">
-              <label htmlFor="email" className="block text-sm">
-                Email address
-              </label>
-              <input
-                type="email"
-                name="email"
-                id="email"
-                required
-                placeholder="Email address"
-                className="w-full px-3 py-2 border rounded-md border-gray-300 bg-gray-50 text-gray-800 focus:border-violet-600"
-              />
-            </div>
-            <div className="space-y-2">
-              <label htmlFor="email" className="block text-sm">
-                photoURL
-              </label>
-              <input
-                type="text"
-                name="photoURL"
-                id="photoURL"
-                required
-                placeholder="photoURL"
-                className="w-full px-3 py-2 border rounded-md border-gray-300 bg-gray-50 text-gray-800 focus:border-violet-600"
-              />
-            </div>
-            <div className="space-y-2">
-              <div className="flex justify-between">
-                <label htmlFor="password" className="text-sm">
-                  Password
-                </label>
-              </div>
-              <div className="relative">
-                <input
-                  type={showPassword ? "text" : "password"}
-                  name="password"
-                  id="password"
-                  placeholder="Password"
-                  className="w-full px-3 py-2 border rounded-md border-gray-300 bg-gray-50 dark:text-gray-800 focus:border-violet-600"
-                />
-                <span
-                  className="absolute right-3 top-3 cursor-pointer text-gray-500"
-                  onClick={togglePasswordVisibility}
-                >
-                  {showPassword ? <EyeOff size={20} /> : <Eye size={20} />}
-                </span>
-              </div>
-            </div>
+        <form onSubmit={handleSubmit(onSubmit)} className="space-y-8">
+          <div>
+            <label className="block text-sm">Name</label>
+            <input type="text" {...register('name', { required: true })} placeholder="Enter Your Name" className="" />
+            {errors.name?.type === 'required' && <p className="text-red-600 text-sm mt-1">Please enter your name</p>}
           </div>
-          <button
-            type="submit"
-            className="w-full px-8 py-3 font-semibold rounded-md bg-[#37b6f5] text-gray-50"
-          >
-            Sign Up
-          </button>
+
+          <div>
+            <label className="block text-sm">Email</label>
+            <input type="email" {...register('email', { required: true })} placeholder="Enter Your Email" className="" />
+            {errors.email?.type === 'required' && <p className="text-red-600 text-sm mt-1">Email is required</p>}
+          </div>
+
+          <div>
+            <label className="block text-sm">Photo</label>
+            <input type="file" onChange={handleImageUp} className="" />
+          </div>
+
+          <div>
+            <label className="block text-sm">Password</label>
+            <input type="password" {...register('password', { required: true })} placeholder="Enter Your password" className="" />
+            {errors.password?.type === "required" && (
+              <p className="text-red-600">Enter a password</p>
+            )}
+            {errors.password?.type === "minLength" && (
+              <p className="text-red-600">Password must be 6+ characters</p>
+            )}
+          </div>
+          <button className=" px-8 py-3 font-semibold rounded-md bg-[#37b6f5] text-gray-50">Sign Up</button>
         </form>
       </div>
       <div className="w-full h-full rounded-md  bg-white mb-1 text-gray-800">
